feat: add CardModel and CardViewCell for tvOS sample app
